requirements.txt
Flask==2.0.1
PyPDF2==3.0.1
Pillow==8.3.1
img2pdf==0.4.4
Werkzeug==2.0.1 